+++ changelog.md
@@ -6,6 +6,8 @@
 - Slight code compression
 - Branch cleanup
 - Readme update
+- Fixed a bug where window would not close on parser failure
+- Fixed parser bug relating to comments
 
 ## 0.0.1
 - Bare bones implementation